pass arethetypeswrong w/ separated d.ts files
