+++ prettier.config.js
@@ -1,7 +1,6 @@
 /** @type {import('prettier').Config} */
 const prettierConfig = {
   singleQuote: true,
-  // plugins: ['prettier-plugin-jsdoc'],
 };
 
 export default prettierConfig;
