+++ scripts/postbuild.ts
@@ -8,7 +8,7 @@
  * by replacing `"/assets/worker.penumbra"` with `"./assets/worker.penumbra"`.
  */
 
-import { readFile, writeFile } from 'node:fs/promises';
+import { readFile, writeFile, copyFile } from 'node:fs/promises';
 import path from 'node:path';
 
 async function patchWebWorkerUrl(
@@ -40,15 +40,35 @@ async function patchWebWorkerUrl(
   }
 }
 
+async function createCtsTypes() {
+  const ctsPath = path.join(process.cwd(), 'dist/main.penumbra.d.cts');
+  const dtsPath = path.join(process.cwd(), 'dist/main.penumbra.d.ts');
+  await copyFile(ctsPath, dtsPath);
+
+  const content = await readFile(ctsPath, 'utf8');
+  const patchedContent = content.replaceAll(
+    / from '(\.\/.*?)'/g,
+    (_, a: string) => ` from '${a}.js'`,
+  );
+  await writeFile(dtsPath, patchedContent, 'utf8');
+
+  console.log(
+    `Successfully copied ${ctsPath} to ${dtsPath} in postbuild script.`,
+  );
+}
+
 async function main() {
   console.group('Postbuild: Patch Web Worker URL');
   try {
     const results = await Promise.allSettled([
       patchWebWorkerUrl('dist/main.penumbra.js'),
       patchWebWorkerUrl('dist/main.penumbra.umd.cjs'),
+      createCtsTypes(),
     ]);
 
-    const errors = results.filter((result) => result.status === 'rejected');
+    const errors = results
+      .filter((result) => result.status === 'rejected')
+      .map((result) => result.reason as unknown);
     if (errors.length > 0) {
       throw new AggregateError(
         errors,
