+++ src/utils/index.ts
@@ -1,12 +1,12 @@
-export { default as emitJobProgress } from './emit-job-progress';
-export { default as emitJobCompletion } from './emit-job-completion';
-export { default as emitZipProgress } from './emit-zip-progress';
-export { default as emitZipCompletion } from './emit-zip-completion';
-export { default as getOrigins } from './get-origins';
-export { default as blobCache } from './blob-cache';
-export { default as isViewableText } from './is-viewable-text';
-export { default as isNumber } from './is-number';
-export { default as throwOutside } from './throw-outside';
-export { default as parseBase64OrUint8Array } from './parse-base64-or-uint8array';
-export { default as emitError } from './emit-error';
-export { default as onPenumbraEvent } from './forward-events';
+export { default as emitJobProgress } from './emit-job-progress.js';
+export { default as emitJobCompletion } from './emit-job-completion.js';
+export { default as emitZipProgress } from './emit-zip-progress.js';
+export { default as emitZipCompletion } from './emit-zip-completion.js';
+export { default as getOrigins } from './get-origins.js';
+export { default as blobCache } from './blob-cache.js';
+export { default as isViewableText } from './is-viewable-text.js';
+export { default as isNumber } from './is-number.js';
+export { default as throwOutside } from './throw-outside.js';
+export { default as parseBase64OrUint8Array } from './parse-base64-or-uint8array.js';
+export { default as emitError } from './emit-error.js';
+export { default as onPenumbraEvent } from './forward-events.js';
